Agregar policy y checks para los sharedPets

+++ mascotas-api/app/Http/Controllers/ContactController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\SharedPets\AcceptRequest;
 use App\Http\Requests\SharedPets\CreateRequest;
 use App\Http\Requests\SharedPets\UpdateRequest;
+use App\Models\SharedPet;
 use App\Repositories\ContactRepository;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Auth;
@@ -20,6 +21,8 @@ class ContactController extends Controller
 
     public function getRequestsByProfessional(): JsonResponse
     {
+        $this->authorize('viewOwnerSharedPets', SharedPet::class);
+
         $professionals = $this->contactRepository->getRequestsByProfessional(Auth::user()->id);
 
         return response()->json([
@@ -30,6 +33,8 @@ class ContactController extends Controller
 
     public function getOwnerSharedPets(): JsonResponse
     {
+        $this->authorize('viewOwnerSharedPets', SharedPet::class);
+
         $sharedPets = $this->contactRepository->getOwnerSharedPets(Auth::user()->id);
 
         return response()->json([
@@ -40,6 +45,8 @@ class ContactController extends Controller
 
     public function getProfessionalSharedPets(): JsonResponse
     {
+        $this->authorize('viewProfessionalSharedPets', SharedPet::class);
+
         $sharedPets = $this->contactRepository->getProfessionalSharedPets(Auth::user()->id);
 
         return response()->json([
@@ -50,6 +57,8 @@ class ContactController extends Controller
 
     public function createSharedPetRequest(CreateRequest $request): JsonResponse
     {
+        $this->authorize('create', SharedPet::class);
+
         $requestCreated = $this->contactRepository->createSharedPetRequest(
             $request->get('description'),
             $request->get('expiration_date'),
@@ -66,6 +75,9 @@ class ContactController extends Controller
 
     public function updateSharedPetRequest(UpdateRequest $request, int $request_id): JsonResponse
     {
+        $request = $this->contactRepository->find($request_id);
+        $this->authorize('update', $request);
+
         $requestUpdated = $this->contactRepository->updateSharedPetRequest(
             $request_id,
             $request->get('description'),
@@ -79,8 +91,11 @@ class ContactController extends Controller
         ]);
     }
 
-    public function acceptSharedPetRequest(AcceptRequest $request, int $request_id): JsonResponse
+    public function acceptSharedPetRequest(int $request_id): JsonResponse
     {
+        $request = $this->contactRepository->find($request_id);
+        $this->authorize('accept', $request);
+
         $request = $this->contactRepository->acceptSharedPetRequest($request_id);
 
         return response()->json([
@@ -89,8 +104,11 @@ class ContactController extends Controller
         ]);
     }
 
-    public function generateAcceptSharedPetRequest(AcceptRequest $request, int $request_id): JsonResponse
+    public function generateAcceptSharedPetRequest(int $request_id): JsonResponse
     {
+        $request = $this->contactRepository->find($request_id);
+        $this->authorize('accept', $request);
+        
         $request = $this->contactRepository->findSharedPetRequest($request_id);
 
         return response()->json([
@@ -99,8 +117,11 @@ class ContactController extends Controller
         ]);
     }
 
-    public function deleteSharedPetRequest(AcceptRequest $request, int $request_id): JsonResponse
+    public function deleteSharedPetRequest(int $request_id): JsonResponse
     {
+        $request = $this->contactRepository->find($request_id);
+        $this->authorize('delete', $request);
+
         $this->contactRepository->deleteSharedPetRequest($request_id);
 
         return response()->json([
